+++ backend/src/config.py
@@ -1,0 +1,2 @@
+DATABASE_URL = "sqlite:///./news_database.db"
+NEWS_API_KEY = "your_news_api_key"
